+++ api/typescript/src/core/lib/enclaves/grpc_node_api_container_client.ts
@@ -162,38 +162,6 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
         return ok(resultExecuteKurtosisModule.value)
     }
 
-    public async executeStartosisRemoteModule(startosisRemoteModuleArgs:ExecuteStartosisModuleArgs): Promise<Result<ExecuteStartosisResponse, Error>> {
-        const promiseExecuteStartosisRemoteModule: Promise<Result<ExecuteStartosisResponse, Error>> = new Promise((resolve, _unusedReject) => {
-            this.client.executeStartosisModule(startosisRemoteModuleArgs, (error: ServiceError | null, response?: ExecuteStartosisResponse) => {
-                if (error === null) {
-                    if (!response) {
-                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
-                    } else {
-                        resolve(ok(response!));
-                    }
-                } else {
-                    resolve(err(error));
-                }
-            })
-        })
-        const resultExecuteStartosisRemoteModule: Result<ExecuteStartosisResponse, Error> = await promiseExecuteStartosisRemoteModule;
-        if (resultExecuteStartosisRemoteModule.isErr()) {
-            return err(resultExecuteStartosisRemoteModule.error)
-        }
-        return ok(resultExecuteStartosisRemoteModule.value)
-    }
-
-    public async executeKurtosisRemoteModule(startosisRemoteModuleArgs:ExecuteStartosisModuleArgs): Promise<Result<Readable, Error>> {
-        const promiseExecuteKurtosisRemoteModule: Promise<Result<ClientReadableStream<KurtosisExecutionResponseLine>, Error>> = new Promise((resolve, _unusedReject) => {
-            resolve(ok(this.client.executeKurtosisModule(startosisRemoteModuleArgs)));
-        })
-        const resultExecuteKurtosisRemoteModule: Result<Readable, Error> = await promiseExecuteKurtosisRemoteModule;
-        if (resultExecuteKurtosisRemoteModule.isErr()) {
-            return err(resultExecuteKurtosisRemoteModule.error)
-        }
-        return ok(resultExecuteKurtosisRemoteModule.value)
-    }
-
     public async startServices(startServicesArgs: StartServicesArgs): Promise<Result<StartServicesResponse, Error>>{
         const promiseStartServices: Promise<Result<StartServicesResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.startServices(startServicesArgs, (error: ServiceError | null, response?: StartServicesResponse) => {