+++ docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Breaking Changes
+- Renamed `kurtosis.mod` file to `kurtosis.yml` this file extension enable syntax highlighting
+  - Users will have to rename all theirs `kurtosis.mod` files
 
 # 0.54.0