+++ docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Changes
+- Small cleanup in `grpc_web_api_container_client` and `grpc_node_api_container_client`. They were implementing executeRemoteKurtosisModule unnecessarily
+
 # 0.54.0
 
 ### Breaking Changes
@@ -10,7 +13,6 @@
 ### Features
 - Updated the CLI to consume the streaming endpoints to execute Startosis. Kurtosis Instructions are now returned live, but the script output is still printed at the end (until we have better formatting).
 - Update integration tests to consume Startosis streaming endpoints
-- Small cleanup in `grpc_web_api_container_client` and `grpc_node_api_container_client`. They were implementing executeRemoteKurtosisModule unnecessarily
 
 # 0.53.12